requirements.txt
pyautogen>=0.2.0
openai>=1.0.0
